src/backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
langchain
langchain-community
langchain-core
langchain-openai
langchain-chroma
langchain-text-splitters
langchain-experimental
openai
pypdf
pypdfium2
pdfminer.six
unstructured
docx2txt
python-dotenv
dill
regex
overrides
wrapt
pymupdf
